Create devtools-environment
