+++ packages/devtools-environment/README.md
@@ -1,0 +1,8 @@
+## Environment
+
+[![Npm version](https://img.shields.io/npm/v/devtools-environment.svg)](https://npmjs.org/package/devtools-environment)
+
+* `isFirefox` is the tool is running in firefox
+* `isFirefoxPanel` - is the tool running in the firefox panel (toolbox)
+* `isDevelopment` - is the tool running in the launchpad
+* `isTesting` - is the tool running in a mochitest environment
